requirements.txt
addict==2.4.0
anyio==4.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
easydict==1.13
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.61.0
fsspec==2025.9.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.2.3
idna==3.4
jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
markupsafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mmcv==2.1.0
mmengine==0.10.7
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==25.0
pillow==11.3.0
platformdirs==4.5.1
protobuf==6.33.1
pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyyaml==6.0.3
regex==2025.11.3
requests==2.28.1
rich==14.2.0
safetensors==0.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
tensorboardx==2.6.4
termcolor==3.2.0
timm==1.0.22
tomli==2.3.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.67.1
triton==2.1.0
typer-slim==0.20.0
typing-extensions==4.15.0
urllib3==1.26.13
wcwidth==0.2.14
yapf==0.43.0
